@@ -1,6 +1,6 @@
 {
   "name": "formsy-react",
-  "version": "2.0.0-beta.4",
+  "version": "2.0.0",
   "description": "A form input builder and validator for React",
   "keywords": [
     "form",
@@ -21,6 +21,11 @@
     }
   },
   "jest": {
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "src/**/*.tsx",
+      "src/**/*.ts"
+    ],
     "transform": {
       "^.+\\.(js|jsx|ts|tsx)$": "babel-jest"
     },
@@ -51,7 +56,7 @@
     "changelog": "auto-changelog",
     "deploy": "np",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,md}\"",
-    "lint": "eslint src/**/*",
+    "lint": "eslint .",
     "preversion": "npm run lint",
     "test": "jest",
     "version": "npm run build && git add dist && npm run changelog && git add CHANGELOG.md"
@@ -69,32 +74,33 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
     "@types/enzyme": "^3.10.3",
-    "@types/jest": "^24.0.17",
+    "@types/jest": "^25.1.1",
     "@types/prop-types": "^15.7.1",
     "@types/react": "^16.8.24",
     "@types/react-dom": "^16.8.5",
-    "@typescript-eslint/eslint-plugin": "^1.13.0",
-    "@typescript-eslint/parser": "^1.13.0",
+    "@typescript-eslint/eslint-plugin": "^2.14.0",
+    "@typescript-eslint/parser": "^2.14.0",
     "auto-changelog": "^1.14.1",
     "babel-eslint": "^10.0.2",
-    "babel-jest": "^24.8.0",
+    "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
     "babelrc-rollup": "^3.0.0",
+    "coveralls": "^3.0.9",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^6.1.0",
-    "eslint-config-airbnb": "^17.1.1",
+    "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^6.0.0",
     "eslint-config-react": "^1.1.7",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.14.3",
-    "husky": "^3.0.2",
-    "jest": "^24.8.0",
-    "np": "^5.0.2",
+    "husky": "^4.2.1",
+    "jest": "^25.1.0",
+    "np": "^5.0.0",
     "prettier": "^1.18.2",
-    "pretty-quick": "^1.11.1",
+    "pretty-quick": "^2.0.0",
     "react": "^16.2.0 || ^16.0.0",
     "react-dom": "^16.2.0 || ^16.0.0",
     "rollup": "^1.19.2",
@@ -102,11 +108,11 @@
     "rollup-plugin-commonjs": "^10.0.2",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-peer-deps-external": "^2.2.0",
-    "sinon": "^7.3.2",
+    "sinon": "^8.0.2",
     "typescript": "^3.5.3"
   },
   "peerDependencies": {
-    "react": "^15.6.1 || ^16.0.0",
-    "react-dom": "^15.6.1 || ^16.0.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   }
 }
