@@ -1,7 +1,6 @@
 import React from 'react';
 import sinon from 'sinon';
 import { mount } from 'enzyme';
-
 import Formsy, { withFormsy } from '../src';
 import immediate from '../__test_utils__/immediate';
 import TestInput, { InputFactory } from '../__test_utils__/TestInput';
@@ -26,14 +25,15 @@ describe('Element', () => {
         updateValue = event => {
           this.props.setValue(event.target.value, false);
         };
+
         render() {
-          return <input type="text" value={this.props.value} onChange={this.updateValue} />;
+          return <input type="text" value={this.props.value || ''} onChange={this.updateValue} />;
         }
       },
     );
     const form = mount(
       <Formsy>
-        <Input name="foo" value="foo" innerRef="comp" />
+        <Input name="foo" value="foo" />
       </Formsy>,
     );
     const inputComponent = form.find(Input);
@@ -49,7 +49,7 @@ describe('Element', () => {
   it('should set back to pristine value when running reset', () => {
     let reset = null;
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         reset = this.props.resetValue;
       },
     });
@@ -68,7 +68,7 @@ describe('Element', () => {
   it('should return error message passed when calling getErrorMessage()', () => {
     let errorMessage = null;
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         errorMessage = this.props.errorMessage;
       },
     });
@@ -84,8 +84,8 @@ describe('Element', () => {
   it('should return true or false when calling isValid() depending on valid state', () => {
     let isValid = null;
     const Input = InputFactory({
-      componentWillReceiveProps: function(nextProps) {
-        isValid = nextProps.isValid;
+      componentDidUpdate: function() {
+        isValid = this.props.isValid;
       },
     });
     const form = mount(
@@ -103,7 +103,7 @@ describe('Element', () => {
   it('should return true or false when calling isRequired() depending on passed required attribute', () => {
     const isRequireds = [];
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         isRequireds.push(this.props.isRequired);
       },
     });
@@ -123,7 +123,7 @@ describe('Element', () => {
   it('should return true or false when calling showRequired() depending on input being empty and required is passed, or not', () => {
     const showRequireds = [];
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         showRequireds.push(this.props.showRequired);
       },
     });
@@ -303,7 +303,7 @@ describe('Element', () => {
     ).toEqual(false);
   });
 
-  it('should not override error messages with error messages passed by form if passed eror messages is an empty object', () => {
+  it('should not override error messages with error messages passed by form if passed error messages is an empty object', () => {
     class TestForm extends React.Component {
       render() {
         return (
@@ -325,6 +325,40 @@ describe('Element', () => {
 
     const inputComponent = form.find(TestInput);
     expect(inputComponent.instance().getErrorMessage()).toEqual('bar3');
+  });
+
+  it('should handle multiple validation error messages passed from validators', () => {
+    function customValidationA() {
+      return 'error message one';
+    }
+
+    function customValidationB() {
+      return 'error message two';
+    }
+
+    function TestForm() {
+      return (
+        <Formsy>
+          <TestInput
+            name="A"
+            validations={{
+              customValidationA,
+              customValidationB,
+            }}
+            value="foo"
+          />
+        </Formsy>
+      );
+    }
+
+    const form = mount(<TestForm />);
+    const inputComponent = form.find(TestInput);
+
+    const formEl = form.find('form');
+    formEl.simulate('submit');
+
+    expect(inputComponent.instance().getErrorMessage()).toEqual('error message one');
+    expect(inputComponent.instance().getErrorMessages()).toEqual(['error message one', 'error message two']);
   });
 
   it('should override all error messages with error messages passed by form', () => {
@@ -512,6 +546,7 @@ describe('Element', () => {
     }
     const form = mount(<TestForm />);
 
+    // TODO: Beef up this smoke test
     expect(true).toBe(true);
 
     const formEl = form.find('form');
@@ -534,6 +569,7 @@ describe('Element', () => {
     }
     const form = mount(<TestForm />);
 
+    // TODO: Beef up this smoke test
     expect(true).toBe(true);
 
     const formEl = form.find('form');
@@ -590,22 +626,16 @@ describe('Element', () => {
     expect(renderSpy.calledTwice).toEqual(true);
   });
 
-  it('binds all necessary methods', () => {
-    const onInputRef = input => {
-      ['isValidValue', 'resetValue', 'setValidations', 'setValue'].forEach(fnName => {
-        const fn = input[fnName];
-        try {
-          fn();
-        } catch (e) {
-          throw new Error(`Method '${fnName}' isn't bound.`);
-        }
-      });
-    };
+  it('unregisters on unmount', () => {
+    const TestComponent = ({ hasInput }) => <Formsy>{hasInput ? <TestInput name="foo" value="foo" /> : null}</Formsy>;
 
-    mount(
-      <Formsy>
-        <TestInput ref={onInputRef} name="name" value="foo" />
-      </Formsy>,
-    );
+    const wrapper = mount(<TestComponent hasInput />);
+    const formsy = wrapper.find(Formsy).instance();
+
+    expect(formsy.inputs).toHaveLength(1);
+
+    wrapper.setProps({ hasInput: false });
+
+    expect(formsy.inputs).toHaveLength(0);
   });
 });
