@@ -1,15 +1,17 @@
-import React from 'react';
+/* eslint-disable max-classes-per-file */
 import { mount } from 'enzyme';
-import sinon from 'sinon';
-
-import Formsy, { addValidationRule } from '../src';
+import * as React from 'react';
+import DynamicInputForm from '../__test_utils__/DynamicInputForm';
 import TestInput from '../__test_utils__/TestInput';
 import TestInputHoc from '../__test_utils__/TestInputHoc';
-import immediate from '../__test_utils__/immediate';
+
+import Formsy, { addValidationRule } from '../src';
 
 describe('Setting up a form', () => {
   it('should expose the users DOM node through an innerRef prop', () => {
     class TestForm extends React.Component {
+      public inputRef: typeof TestInputHoc;
+
       render() {
         return (
           <Formsy>
@@ -25,135 +27,157 @@ describe('Setting up a form', () => {
     }
 
     const form = mount(<TestForm />);
-    const input = form.instance().inputRef;
+    const input = (form.instance() as TestForm).inputRef;
     expect(input.props.name).toEqual('name');
   });
 
   it('should render a form into the document', () => {
-    const form = mount(<Formsy></Formsy>);
+    const form = mount(<Formsy />);
     expect(form.find('form').name()).toEqual('form');
   });
 
   it('should set a class name if passed', () => {
-    const form = mount(<Formsy className="foo"></Formsy>);
+    const form = mount(<Formsy className="foo" />);
     expect(form.find('form').hasClass('foo')).toBe(true);
   });
 
   it('should allow for null/undefined children', () => {
     let model = null;
-    class TestForm extends React.Component {
+
+    function TestForm() {
+      return (
+        <Formsy
+          onSubmit={formModel => {
+            model = formModel;
+          }}
+        >
+          <h1>Test</h1>
+          {null}
+          {undefined}
+          <TestInput name="name" value="foo" />
+        </Formsy>
+      );
+    }
+
+    const form = mount(<TestForm />);
+    form.simulate('submit');
+    expect(model).toEqual({ name: 'foo' });
+  });
+
+  it('should allow for inputs being added dynamically', () => {
+    let model = null;
+
+    const form = mount(
+      <DynamicInputForm
+        onSubmit={formModel => {
+          model = formModel;
+        }}
+        inputName="test"
+      />,
+    );
+    form.find('button').simulate('click');
+    form.update();
+
+    form.simulate('submit');
+    expect(model).toHaveProperty('test');
+  });
+
+  it('should allow dynamically added inputs to update the form-model', () => {
+    let model = null;
+
+    const form = mount(
+      <DynamicInputForm
+        onSubmit={formModel => {
+          model = formModel;
+        }}
+        inputName="test"
+      />,
+    );
+    form.find('button').simulate('click');
+    form.update();
+
+    form.find('input').simulate('change', {
+      target: { value: 'foo' },
+    });
+    form.simulate('submit');
+    expect(model).toHaveProperty('test', 'foo');
+  });
+
+  it('should allow a dynamically updated input to update the form-model', () => {
+    let model = null;
+
+    class TestForm extends React.Component<{ inputValue: any }, { inputValue: any }> {
+      constructor(props) {
+        super(props);
+        this.state = { inputValue: props.inputValue };
+      }
+
+      updateInputValue = () => this.setState({ inputValue: 'bar' });
+
       render() {
+        const { inputValue } = this.state;
         return (
-          <Formsy onSubmit={formModel => (model = formModel)}>
-            <h1>Test</h1>
-            {null}
-            {undefined}
-            <TestInput name="name" value={'foo'} />
+          <Formsy
+            onSubmit={formModel => {
+              model = formModel;
+            }}
+          >
+            <TestInput name="test" value={inputValue} />
+            <button type="button" onClick={this.updateInputValue}>
+              Update
+            </button>
           </Formsy>
         );
       }
     }
 
-    const form = mount(<TestForm />);
-    immediate(() => {
-      expect(model).toEqual({ name: 'foo' });
-    });
+    const form = mount(<TestForm inputValue="foo" />);
+
+    form.simulate('submit');
+
+    expect(model).toHaveProperty('test', 'foo');
+
+    form.find('button').simulate('click');
+    form.update();
+    form.simulate('submit');
+    form.update();
+
+    expect(model).toHaveProperty('test', 'bar');
   });
+});
 
-  it('should allow for inputs being added dynamically', () => {
-    const inputs = [];
-    let forceUpdate = null;
-    let model = null;
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
-      render() {
-        return <Formsy onSubmit={formModel => (model = formModel)}>{inputs}</Formsy>;
-      }
+describe('mapModel', () => {
+  it('should honor mapModel transformations', () => {
+    const mapping = jest.fn(model => ({
+      ...model,
+      testChange: true,
+    }));
+    const onSubmit = jest.fn();
+
+    function TestForm() {
+      return (
+        <Formsy mapping={mapping} onSubmit={onSubmit}>
+          <TestInput name="parent.child" value="test" />
+        </Formsy>
+      );
     }
-    const form = mount(<TestForm />);
 
-    // Wait before adding the input
-    setTimeout(() => {
-      inputs.push(<TestInput name="test" value="" key={inputs.length} />);
-
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.simulate('submit');
-          test.ok('test' in model);
-        });
-      });
-    }, 10);
-  });
-
-  it('should allow dynamically added inputs to update the form-model', () => {
-    const inputs = [];
-    let forceUpdate = null;
-    let model = null;
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
-      render() {
-        return <Formsy onSubmit={formModel => (model = formModel)}>{inputs}</Formsy>;
-      }
-    }
     const form = mount(<TestForm />);
 
-    // Wait before adding the input
-    immediate(() => {
-      inputs.push(<TestInput name="test" key={inputs.length} />);
-
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.find('input').simulate({
-            target: { value: 'foo' },
-          });
-          form.simulate('submit');
-          expect(model.test).toEqual('foo');
-        });
-      });
-    });
-  });
-
-  it('should allow a dynamically updated input to update the form-model', () => {
-    let forceUpdate = null;
-    let model = null;
-
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
-      render() {
-        const input = <TestInput name="test" value={this.props.value} />;
-
-        return <Formsy onSubmit={formModel => (model = formModel)}>{input}</Formsy>;
-      }
-    }
-    let form = mount(<TestForm value="foo" />);
-
-    // Wait before changing the input
-    immediate(() => {
-      form = mount(<TestForm value="bar" />);
-
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.simulate('submit');
-          expect(model.test).toEqual('bar');
-        });
-      });
-    });
+    form.simulate('submit');
+    expect(mapping).toHaveBeenCalledWith({ 'parent.child': 'test' });
+    expect(onSubmit).toHaveBeenCalledWith(
+      { 'parent.child': 'test', testChange: true },
+      expect.any(Function),
+      expect.any(Function),
+    );
   });
 });
 
 describe('validations', () => {
   it('should run when the input changes', () => {
-    const runRule = sinon.spy();
-    const notRunRule = sinon.spy();
+    const runRule = jest.fn();
+    const notRunRule = jest.fn();
 
     addValidationRule('runRule', runRule);
     addValidationRule('notRunRule', notRunRule);
@@ -168,13 +192,14 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(runRule.calledWith({ one: 'bar' })).toBe(true);
-    expect(notRunRule.called).toEqual(false);
+
+    expect(runRule).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
+    expect(notRunRule).not.toHaveBeenCalled();
   });
 
   it('should allow the validation to be changed', () => {
-    const ruleA = sinon.spy();
-    const ruleB = sinon.spy();
+    const ruleA = jest.fn();
+    const ruleB = jest.fn();
     addValidationRule('ruleA', ruleA);
     addValidationRule('ruleB', ruleB);
 
@@ -183,11 +208,13 @@ describe('validations', () => {
         super(props);
         this.state = { rule: 'ruleA' };
       }
+
       changeRule = () => {
         this.setState({
           rule: 'ruleB',
         });
       };
+
       render() {
         return (
           <Formsy>
@@ -203,25 +230,29 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(ruleB.calledWith({ one: 'bar' })).toBe(true);
+    expect(ruleB).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
   });
 
   it('should invalidate a form if dynamically inserted input is invalid', () => {
-    const isInValidSpy = sinon.spy();
+    const isInValidSpy = jest.fn();
 
     class TestForm extends React.Component {
+      formRef = React.createRef();
+
       constructor(props) {
         super(props);
         this.state = { showSecondInput: false };
       }
+
       addInput = () => {
         this.setState({
           showSecondInput: true,
         });
       };
+
       render() {
         return (
-          <Formsy ref="formsy" onInvalid={isInValidSpy}>
+          <Formsy ref={this.formRef} onInvalid={isInValidSpy}>
             <TestInput name="one" validations="isEmail" value="foo@bar.com" />
             {this.state.showSecondInput ? <TestInput name="two" validations="isEmail" value="foo@bar" /> : null}
           </Formsy>
@@ -231,30 +262,32 @@ describe('validations', () => {
 
     const form = mount(<TestForm />);
 
-    expect(form.instance().refs.formsy.state.isValid).toEqual(true);
+    expect(form.instance().formRef.current.state.isValid).toEqual(true);
     form.instance().addInput();
 
-    immediate(() => {
-      expect(isInValidSpy.called).toEqual(true);
-    });
+    expect(isInValidSpy).toHaveBeenCalled();
   });
 
   it('should validate a form when removing an invalid input', () => {
-    const isValidSpy = sinon.spy();
+    const isValidSpy = jest.fn();
 
     class TestForm extends React.Component {
+      formRef = React.createRef();
+
       constructor(props) {
         super(props);
         this.state = { showSecondInput: true };
       }
+
       removeInput() {
         this.setState({
           showSecondInput: false,
         });
       }
+
       render() {
         return (
-          <Formsy ref="formsy" onValid={isValidSpy}>
+          <Formsy ref={this.formRef} onValid={isValidSpy}>
             <TestInput name="one" validations="isEmail" value="foo@bar.com" />
             {this.state.showSecondInput ? <TestInput name="two" validations="isEmail" value="foo@bar" /> : null}
           </Formsy>
@@ -264,17 +297,15 @@ describe('validations', () => {
 
     const form = mount(<TestForm />);
 
-    expect(form.instance().refs.formsy.state.isValid).toEqual(false);
+    expect(form.instance().formRef.current.state.isValid).toEqual(false);
     form.instance().removeInput();
 
-    immediate(() => {
-      expect(isValidSpy.called).toEqual(true);
-    });
+    expect(isValidSpy).toHaveBeenCalled();
   });
 
   it('runs multiple validations', () => {
-    const ruleA = sinon.spy();
-    const ruleB = sinon.spy();
+    const ruleA = jest.fn();
+    const ruleB = jest.fn();
     addValidationRule('ruleA', ruleA);
     addValidationRule('ruleB', ruleB);
 
@@ -288,45 +319,50 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(ruleA.calledWith({ one: 'bar' })).toBe(true);
-    expect(ruleB.calledWith({ one: 'bar' })).toBe(true);
+    expect(ruleA).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
+    expect(ruleB).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
   });
 });
 
 describe('onChange', () => {
   it('should not trigger onChange when form is mounted', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       render() {
-        return <Formsy onChange={hasChanged}></Formsy>;
+        return <Formsy onChange={hasChanged} />;
       }
     }
+
     mount(<TestForm />);
-    expect(hasChanged.called).toEqual(false);
+    expect(hasChanged).not.toHaveBeenCalled();
   });
 
   it('should trigger onChange once when form element is changed', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasChanged}>
-        <TestInput name="foo" />
+        <TestInput name="foo" value="" />
       </Formsy>,
     );
     form.find('input').simulate('change', { target: { value: 'bar' } });
-    expect(hasChanged.calledOnce).toEqual(true);
+    expect(hasChanged).toHaveBeenCalledTimes(1);
   });
 
   it('should trigger onChange once when new input is added to form', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         showInput: false,
       };
+
       addInput() {
         this.setState({
           showInput: true,
         });
       }
+
       render() {
         return <Formsy onChange={hasChanged}>{this.state.showInput ? <TestInput name="test" /> : null}</Formsy>;
       }
@@ -334,22 +370,23 @@ describe('onChange', () => {
 
     const form = mount(<TestForm />);
     form.instance().addInput();
-    immediate(() => {
-      expect(hasChanged.calledOnce).toEqual(true);
-    });
+    expect(hasChanged).toHaveBeenCalledTimes(1);
   });
 
   it('onChange should honor dot notation transformations', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         showInput: false,
       };
+
       addInput() {
         this.setState({
           showInput: true,
         });
       }
+
       render() {
         return (
           <Formsy onChange={hasChanged}>
@@ -361,9 +398,9 @@ describe('onChange', () => {
 
     const form = mount(<TestForm />);
     form.instance().addInput();
-    immediate(() => {
-      expect(hasChanged.args[0][0]).toEqual({ parent: { child: 'test' } });
-    });
+    form.update();
+
+    expect(hasChanged).toHaveBeenCalledWith({ parent: { child: 'test' } }, false);
   });
 });
 
@@ -371,9 +408,11 @@ describe('Update a form', () => {
   it('should allow elements to check if the form is disabled', () => {
     class TestForm extends React.Component {
       state = { disabled: true };
+
       enableForm() {
         this.setState({ disabled: false });
       }
+
       render() {
         return (
           <Formsy disabled={this.state.disabled}>
@@ -388,38 +427,34 @@ describe('Update a form', () => {
     expect(input.instance().isFormDisabled()).toEqual(true);
 
     form.instance().enableForm();
-    immediate(() => {
-      expect(input.instance().isFormDisabled()).toEqual(false);
-    });
+
+    expect(input.instance().isFormDisabled()).toEqual(false);
   });
 
   it('should be possible to pass error state of elements by changing an errors attribute', () => {
     class TestForm extends React.Component {
       state = { validationErrors: { foo: 'bar' } };
+
       onChange = values => {
         this.setState(values.foo ? { validationErrors: {} } : { validationErrors: { foo: 'bar' } });
       };
+
       render() {
         return (
           <Formsy onChange={this.onChange} validationErrors={this.state.validationErrors}>
-            <TestInput name="foo" />
+            <TestInput name="foo" value="" />
           </Formsy>
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      const input = form.find(TestInput);
-      expect(input.instance().getErrorMessage()).toEqual('bar');
-      input.setValue('gotValue');
+    const input = form.find(TestInput);
+    expect(input.instance().getErrorMessage()).toEqual('bar');
+    input.instance().setValue('gotValue');
 
-      // Wait for update
-      immediate(() => {
-        expect(input.instance().getErrorMessage()).toEqual(null);
-      });
-    });
+    expect(input.instance().getErrorMessage()).toEqual(null);
   });
 
   it('should prevent a default submit', () => {
@@ -458,7 +493,8 @@ describe('Update a form', () => {
 
 
   it('should trigger an onValidSubmit when submitting a valid form', () => {
-    let isCalled = sinon.spy();
+    const isCalled = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -468,14 +504,16 @@ describe('Update a form', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const FoundForm = form.find(TestForm);
     FoundForm.simulate('submit');
-    expect(isCalled.called).toEqual(true);
+    expect(isCalled).toHaveBeenCalled();
   });
 
   it('should trigger an onInvalidSubmit when submitting an invalid form', () => {
-    let isCalled = sinon.spy();
+    const isCalled = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -485,17 +523,19 @@ describe('Update a form', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
     const FoundForm = form.find(TestForm);
     FoundForm.simulate('submit');
-    expect(isCalled.called).toEqual(true);
+    expect(isCalled).toHaveBeenCalled();
   });
 });
 
 describe('value === false', () => {
   it('should call onSubmit correctly', () => {
-    const onSubmit = sinon.spy();
+    const onSubmit = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -509,20 +549,23 @@ describe('value === false', () => {
 
     const form = mount(<TestForm />);
     form.simulate('submit');
-    expect(onSubmit.calledWith({ foo: false })).toEqual(true);
+    expect(onSubmit).toHaveBeenCalledWith({ foo: false }, expect.any(Function), expect.any(Function));
   });
 
   it('should allow dynamic changes to false', () => {
-    const onSubmit = sinon.spy();
+    const onSubmit = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy onSubmit={onSubmit}>
@@ -536,7 +579,7 @@ describe('value === false', () => {
     const form = mount(<TestForm />);
     form.instance().changeValue();
     form.simulate('submit');
-    expect(onSubmit.calledWith({ foo: false })).toEqual(true);
+    expect(onSubmit).toHaveBeenCalledWith({ foo: false }, expect.any(Function), expect.any(Function));
   });
 
   it('should say the form is submitted', () => {
@@ -544,12 +587,13 @@ describe('value === false', () => {
       render() {
         return (
           <Formsy>
-            <TestInput name="foo" value={true} type="checkbox" />
+            <TestInput name="foo" value type="checkbox" />
             <button type="submit">Save</button>
           </Formsy>
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     expect(input.instance().isFormSubmitted()).toEqual(false);
@@ -562,11 +606,13 @@ describe('value === false', () => {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy>
@@ -576,6 +622,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -586,12 +633,13 @@ describe('value === false', () => {
     expect(input.instance().getValue()).toEqual(true);
   });
 
-  it('should be able to set a value to one component with setValue', () => {
+  it('should be able to set a value to components with updateInputsWithValue', () => {
     class TestForm extends React.Component {
       state = {
         valueFoo: true,
         valueBar: true,
       };
+
       render() {
         return (
           <Formsy>
@@ -602,6 +650,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const inputs = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -617,7 +666,7 @@ describe('value === false', () => {
         .instance()
         .getValue(),
     ).toEqual(true);
-    formsyForm.instance().setValue('foo', false);
+    formsyForm.instance().updateInputsWithValue({ foo: false });
     expect(
       inputs
         .at(0)
@@ -637,11 +686,13 @@ describe('value === false', () => {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy>
@@ -651,6 +702,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -677,6 +729,7 @@ describe('.reset()', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -686,22 +739,43 @@ describe('.reset()', () => {
     });
     expect(input.instance().getValue()).toEqual('');
   });
+
+  it('should be able to reset the form using a button', () => {
+    function TestForm() {
+      return (
+        <Formsy>
+          <TestInput name="foo" value="foo" />
+          <button type="submit">Save</button>
+        </Formsy>
+      );
+    }
+
+    const form = mount(<TestForm />);
+    const input = form.find(TestInput);
+    const formsyForm = form.find(Formsy);
+
+    expect(input.instance().getValue()).toEqual('foo');
+    input.simulate('change', { target: { value: 'foobar' } });
+    expect(input.instance().getValue()).toEqual('foobar');
+    formsyForm.simulate('reset');
+    expect(input.instance().getValue()).toEqual('foo');
+  });
 });
 
 describe('.isChanged()', () => {
   it('initially returns false', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
       </Formsy>,
     );
     expect(form.instance().isChanged()).toEqual(false);
-    expect(hasOnChanged.called).toEqual(false);
+    expect(hasOnChanged).not.toHaveBeenCalled();
   });
 
   it('returns true when changed', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
@@ -712,11 +786,11 @@ describe('.isChanged()', () => {
       target: { value: 'bar' },
     });
     expect(form.instance().isChanged()).toEqual(true);
-    expect(hasOnChanged.calledWith({ one: 'bar' })).toEqual(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'bar' }, true);
   });
 
   it('returns false if changes are undone', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
@@ -726,29 +800,34 @@ describe('.isChanged()', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(hasOnChanged.calledWith({ one: 'bar' })).toBe(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'bar' }, true);
 
     input.simulate('change', {
       target: { value: 'foo' },
     });
+
     expect(form.instance().isChanged()).toEqual(false);
-    expect(hasOnChanged.calledWith({ one: 'foo' })).toBe(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'foo' }, false);
   });
 });
 
 describe('form valid state', () => {
   it('should allow to be changed with updateInputsWithError', () => {
+    let isValid = true;
+
     class TestForm extends React.Component {
-      state = { isValid: true };
       onValidSubmit = (model, reset, updateInputsWithError) => {
         updateInputsWithError({ foo: 'bar' }, true);
       };
+
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
+
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy onInvalid={this.onInvalid} onValid={this.onValid} onValidSubmit={this.onValidSubmit}>
@@ -757,32 +836,58 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.simulate('submit');
+    expect(isValid).toEqual(true);
+    form.simulate('submit');
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(false);
-      });
+    expect(isValid).toEqual(false);
+  });
+
+  it('should throw an error when updateInputsWithError is called with a missing input', () => {
+    const mockConsoleError = jest.spyOn(console, 'error');
+    mockConsoleError.mockImplementation(() => {
+      // do nothing
     });
+
+    class TestForm extends React.Component {
+      onValidSubmit = (model, reset, updateInputsWithError) => {
+        updateInputsWithError({ bar: 'bar' }, true);
+      };
+
+      render() {
+        return (
+          <Formsy onValidSubmit={this.onValidSubmit}>
+            <TestInput name="foo" />
+          </Formsy>
+        );
+      }
+    }
+
+    const form = mount(<TestForm />);
+    expect(() => form.simulate('submit')).toThrow();
+    mockConsoleError.mockRestore();
   });
 
   it('should be false when validationErrors is not empty', () => {
+    let isValid = true;
+
     class TestForm extends React.Component {
-      state = { validationErrors: {}, isValid: true };
+      state = { validationErrors: {} };
+
       setValidationErrors = empty => {
         this.setState(!empty ? { validationErrors: { foo: 'bar' } } : { validationErrors: {} });
       };
+
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
+
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy onInvalid={this.onInvalid} onValid={this.onValid} validationErrors={this.state.validationErrors}>
@@ -791,32 +896,32 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.setValidationErrors();
+    expect(isValid).toEqual(true);
+    form.instance().setValidationErrors();
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(false);
-      });
-    });
+    expect(isValid).toEqual(false);
   });
 
   it('should be true when validationErrors is not empty and preventExternalInvalidation is true', () => {
+    let isValid = true;
     class TestForm extends React.Component {
-      state = { validationErrors: {}, isValid: true };
+      state = { validationErrors: {} };
+
       setValidationErrors = empty => {
         this.setState(!empty ? { validationErrors: { foo: 'bar' } } : { validationErrors: {} });
       };
+
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
+
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy
@@ -830,17 +935,12 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.setValidationErrors();
+    expect(isValid).toEqual(true);
+    form.instance().setValidationErrors();
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(true);
-      });
-    });
+    expect(isValid).toEqual(true);
   });
 });
