@@ -1,59 +1,29 @@
 import React from 'react';
-import { mount } from 'enzyme';
 
-import Formsy from '../src';
-import { InputFactory } from '../__test_utils__/TestInput';
-
-const TestInput = InputFactory({
-  render: function() {
-    return <input value={this.props.value} readOnly />;
-  },
-});
-
-class TestForm extends React.Component {
-  render() {
-    return (
-      <Formsy>
-        <TestInput name="foo" validations="equals:foo" value={this.props.inputValue} />
-      </Formsy>
-    );
-  }
-}
+import ValidationForm, { expectIsValid } from '../__test_utils__/expectIsValid';
 
 describe('equals', () => {
   it('should pass when the value is equal', () => {
-    const form = mount(<TestForm inputValue="foo" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="equals:foo" value="foo" />).toEqual(true);
   });
 
   it('should fail when the value is not equal', () => {
-    const form = mount(<TestForm inputValue="fo" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="equals:foo" value="fo" />).toEqual(false);
   });
 
   it('should pass with an empty string', () => {
-    const form = mount(<TestForm inputValue={''} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="equals:foo" value="" />).toEqual(true);
   });
 
   it('should pass with an undefined', () => {
-    const form = mount(<TestForm inputValue={undefined} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="equals:foo" value={undefined} />).toEqual(true);
   });
 
   it('should pass with a null', () => {
-    const form = mount(<TestForm inputValue={null} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="equals:foo" value={null} />).toEqual(true);
   });
 
   it('should fail with a number', () => {
-    const form = mount(<TestForm inputValue={42} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="equals:foo" value={42} />).toEqual(false);
   });
 });
