@@ -1,65 +1,33 @@
 import React from 'react';
-import { mount } from 'enzyme';
 
-import Formsy from '../src';
-import { InputFactory } from '../__test_utils__/TestInput';
-
-const TestInput = InputFactory({
-  render: function() {
-    return <input value={this.props.value} readOnly />;
-  },
-});
-
-class TestForm extends React.Component {
-  render() {
-    return (
-      <Formsy>
-        <TestInput name="foo" validations="isAlpha" value={this.props.inputValue} />
-      </Formsy>
-    );
-  }
-}
+import ValidationForm, { expectIsValid } from '../__test_utils__/expectIsValid';
 
 describe('isAlpha', () => {
   it('should pass with a default value', () => {
-    const form = mount(<TestForm />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isAlpha" />).toEqual(true);
   });
 
   it('should pass with a string is only latin letters', () => {
-    const form = mount(<TestForm inputValue="myValue" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isAlpha" value="myValue" />).toEqual(true);
   });
 
   it('should fail with a string with numbers', () => {
-    const form = mount(<TestForm inputValue="myValue42" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isAlpha" value="myValue42" />).toEqual(false);
   });
 
   it('should pass with an undefined', () => {
-    const form = mount(<TestForm inputValue={undefined} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isAlpha" value={undefined} />).toEqual(true);
   });
 
   it('should pass with a null', () => {
-    const form = mount(<TestForm inputValue={null} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isAlpha" value={null} />).toEqual(true);
   });
 
   it('should pass with an empty string', () => {
-    const form = mount(<TestForm inputValue={''} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isAlpha" value="" />).toEqual(true);
   });
 
   it('should fail with a number', () => {
-    const form = mount(<TestForm inputValue={42} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isAlpha" value={42} />).toEqual(false);
   });
 });
