@@ -1,89 +1,49 @@
 import React from 'react';
-import { mount } from 'enzyme';
 
-import Formsy from '../src';
-import { InputFactory } from '../__test_utils__/TestInput';
-
-const TestInput = InputFactory({
-  render() {
-    return <input value={this.props.value} readOnly />;
-  },
-});
-
-class TestForm extends React.Component {
-  render() {
-    return (
-      <Formsy>
-        <TestInput name="foo" validations="isInt" value={this.props.inputValue} />
-      </Formsy>
-    );
-  }
-}
+import ValidationForm, { expectIsValid } from '../__test_utils__/expectIsValid';
 
 describe('isInt', () => {
   it('should pass with a default value', () => {
-    const form = mount(<TestForm />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" />).toEqual(true);
   });
 
   it('should pass with an empty string', () => {
-    const form = mount(<TestForm inputValue="" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value="" />).toEqual(true);
   });
 
   it('should fail with a string', () => {
-    const form = mount(<TestForm inputValue="abc" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isInt" value="abc" />).toEqual(false);
   });
 
   it('should pass with a number as string', () => {
-    const form = mount(<TestForm inputValue="+42" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value="+42" />).toEqual(true);
   });
 
   it('should fail string with digits', () => {
-    const form = mount(<TestForm inputValue="42 is an answer" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isInt" value="42 is an answer" />).toEqual(false);
   });
 
   it('should pass with an int', () => {
-    const form = mount(<TestForm inputValue={42} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value={42} />).toEqual(true);
   });
 
   it('should fail with a float', () => {
-    const form = mount(<TestForm inputValue={Math.PI} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isInt" value={Math.PI} />).toEqual(false);
   });
 
   it('should fail with a float in science notation', () => {
-    const form = mount(<TestForm inputValue="-1e3" />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(false);
+    expectIsValid(<ValidationForm validations="isInt" value="-1e3" />).toEqual(false);
   });
 
   it('should pass with undefined', () => {
-    const form = mount(<TestForm inputValue={undefined} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value={undefined} />).toEqual(true);
   });
 
   it('should pass with null', () => {
-    const form = mount(<TestForm inputValue={null} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value={null} />).toEqual(true);
   });
 
   it('should pass with a zero', () => {
-    const form = mount(<TestForm inputValue={0} />);
-    const inputComponent = form.find(TestInput);
-    expect(inputComponent.instance().isValid()).toEqual(true);
+    expectIsValid(<ValidationForm validations="isInt" value={0} />).toEqual(true);
   });
 });
